+++ sdk/typescript/test-suite/package.json
@@ -1,17 +1,16 @@
 {
-  "name": "@roochnetwork/rooch-sdk",
+  "name": "@roochnetwork/test-suite",
   "author": "Rooch.network <opensource@rooch.network>",
-  "version": "0.2.4",
-  "description": "Rooch SDK",
+  "version": "0.1.1",
+  "description": "Rooch Test Suite",
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=18.0.0"
-  },
+  "node": ">=18.0.0"
+},
   "packageManager": "pnpm@8.6.6",
   "scripts": {
-    "clean": "rimraf tsconfig.tsbuildinfo rimraf dist",
-    "codegen:version": "node genversion.mjs",
-    "build": "node genversion.mjs && pnpm build:package",
+  "clean": "rimraf tsconfig.tsbuildinfo rimraf dist",
+    "build": "pnpm build:package",
     "build:package": "build-package",
     "vitest": "vitest",
     "test": "pnpm test:unit && pnpm test:e2e",
@@ -28,53 +27,36 @@
     "prettier:check": "prettier -c --ignore-unknown .",
     "prettier:fix": "prettier -w --ignore-unknown .",
     "lint": "pnpm eslint:check && pnpm prettier:check",
-    "lint:fix": "pnpm eslint:fix && pnpm prettier:fix",
-    "gen": "pnpm exec tsx ./scripts/gen.ts && pnpm lint:fix"
-  },
+    "lint:fix": "pnpm eslint:fix && pnpm prettier:fix"
+},
   "repository": {
-    "type": "git",
+  "type": "git",
     "url": "https://github.com/rooch-network/rooch.git"
-  },
+},
   "homepage": "https://github.com/rooch-network/rooch",
   "bugs": {
-    "url": "https://github.com/rooch-network/rooch/issues"
-  },
+  "url": "https://github.com/rooch-network/rooch/issues"
+},
   "files": ["dist", "src"],
   "type": "module",
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
   "types": "./dist/cjs/index.d.ts",
-  "keywords": ["Rooch", "Rooch Network", "Move"],
   "devDependencies": {
-    "@roochnetwork/build-scripts": "workspace:*",
-    "@roochnetwork/test-suite": "workspace:*",
-    "@iarna/toml": "^2.2.5",
+  "@roochnetwork/build-scripts": "workspace:*",
     "@types/node": "^20.14.10",
     "@types/tmp": "^0.2.1",
-    "ts-retry-promise": "^0.7.0",
     "typescript": "^5.3.3",
     "vite": "^4.4.4",
     "vitest": "^1.6.0",
-    "wait-on": "^7.0.1",
     "tmp": "^0.2.1",
-    "@types/bs58check": "2.1.2",
     "testcontainers": "10.11.0"
-  },
+},
   "dependencies": {
-    "@noble/hashes": "~1.4.0",
-    "@noble/curves": "~1.4.0",
-    "@scure/base": "~1.1.6",
-    "@scure/bip32": "^1.3.1",
-    "@scure/bip39": "^1.2.1",
-    "@suchipi/femver": "^1.0.0",
-    "tweetnacl": "^1.0.3",
-    "bech32": "^2.0.0",
-    "@mysten/bcs": "1.0.2",
-    "valibot": "^0.25.0",
-    "bs58check": "2.1.2"
-  },
+
+},
   "publishConfig": {
-    "access": "public",
+  "access": "public",
     "registry": "https://registry.npmjs.org/"
-  }
+}
 }
