+++ sdk/typescript/test-suite/vite.config.ts
@@ -1,0 +1,22 @@
+// Copyright (c) RoochNetwork
+// SPDX-License-Identifier: Apache-2.0
+
+import * as path from 'path'
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+    },
+  },
+  test: {
+    minThreads: 1,
+    maxThreads: 8,
+    hookTimeout: 1000000,
+    testTimeout: 1000000,
+    env: {
+      NODE_ENV: 'test',
+    },
+  },
+})
